requirements.txt
aliyun-python-sdk-alidns==2.6.29
aliyun-python-sdk-core==2.13.35
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Deprecated==1.2.12
DingtalkChatbot==1.5.2
Django==3.1.13
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.4.5
django-filter==2.4.0
django-filter-stubs==0.1.1
django-guardian==2.4.0
django-oauth-toolkit==1.3.0
django-redis==5.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-guardian==0.3.0
djangorestframework-stubs==1.4.0
idna==3.2
importlib-metadata==4.6.3
jmespath==0.10.0
jwcrypto==1.0
kombu==5.1.0
mysql-connector-python==8.0.26
mysqlclient==2.0.3
oauthlib==3.1.1
prompt-toolkit==3.0.19
protobuf==3.17.3
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.5.0
